osmApi: refactor – break up `osmApi.ts`
