+++ src/components/SearchBox/onSelectedFactory.ts
@@ -1,6 +1,6 @@
 import Router, { NextRouter } from 'next/router';
 import { getApiId, getShortId, getUrlOsmId } from '../../services/helpers';
-import { addFeatureCenterToCache } from '../../services/osm/osmApi';
+import { addFeatureCenterToCache } from '../../services/osm/featureCenterToCache';
 import { getOverpassSource } from '../../services/mapStorage';
 import { performOverpassSearch } from '../../services/overpass/overpassSearch';
 import { t } from '../../services/intl';
