+++ src/services/osm/insertOsmNote.ts
@@ -1,0 +1,29 @@
+import { Position, SuccessInfo } from '../types';
+import { fetchJson } from '../fetch';
+import { API_SERVER, OSM_WEBSITE } from './consts';
+
+export const insertOsmNote = async (
+  point: Position,
+  text: string,
+): Promise<SuccessInfo> => {
+  const [lon, lat] = point;
+
+  const body = new URLSearchParams();
+  body.append('lat', `${lat}`);
+  body.append('lon', `${lon}`);
+  body.append('text', text);
+
+  // {"type":"Feature","geometry":{"type":"Point","coordinates":[14.3244982,50.0927863]},"properties":{"id":26569,"url":"https://master.apis.dev.openstreetmap.org/api/0.6/notes/26569.json","comment_url":"https://master.apis.dev.openstreetmap.org/api/0.6/notes/26569/comment.json","close_url":"https://master.apis.dev.openstreetmap.org/api/0.6/notes/26569/close.json","date_created":"2021-04-17 10:37:44 UTC","status":"open","comments":[{"date":"2021-04-17 10:37:44 UTC","action":"opened","text":"way/39695868! Place was marked permanently closed.From https://osmapp.org/way/39695868","html":"\u003cp\u003eway/39695868! Place was marked permanently closed.From \u003ca href=\"https://osmapp.org/way/39695868\" rel=\"nofollow noopener noreferrer\"\u003ehttps://osmapp.org/way/39695868\u003c/a\u003e\u003c/p\u003e"}]}}
+  const reply = await fetchJson(`${API_SERVER}/api/0.6/notes.json`, {
+    nocache: true,
+    method: 'POST',
+    body,
+  });
+
+  const noteId = reply.properties.id;
+  return {
+    type: 'note',
+    text,
+    url: `${OSM_WEBSITE}/note/${noteId}`,
+  };
+};
